Get gulpjs working
GulpJS is working as a build system, however not everything
from Grunt.js made it over.  GulpJS, in my opinion, isn't quite
ready for prime time yet.  A lot of the plugins don't work, are buggy,
seriously lack documentation, etc.

This branch also breaks the Grunt.js build functionality, most likely
because of all of the changes made to Browserify.  The Grunt.js stuff
will need to be fixed on this branch before it can be merged to master.